Replaced "The Shriven" with "His pursuers"
The Shriven are a specific type of character unique to the WarHammer 40K creative universe. While the general concept of certain fantastical characters such as elves and goblins are "generic" enough to not be monopolized through copyright, the detailed nature of the Shriven and its belonging to the WarHammer 40K universe makes me hesitant to incorporate them in this work. Words changed to avoid this potential issue.